requirement.txt
numpy
pandas 
seaborn
matplotlib
scikit-learn
pymongo
python-dotenv
pyYAML
dill
-e .